moved test function, removed pep8 lint
