+++ examples/performance_benchmarks.py
@@ -17,60 +17,9 @@ import time
 from openfermion.ops import (FermionOperator,
                              InteractionOperator,
                              normal_ordered)
+from openfermion.tests._testing_utils import random_interaction_operator
 from openfermion.transforms import get_fermion_operator, jordan_wigner
 from openfermion.utils import jordan_wigner_sparse
-
-
-def artificial_molecular_operator(n_qubits):
-    """Make an artificial random InteractionOperator for testing purposes."""
-
-    # Initialize.
-    constant = numpy.random.randn()
-    one_body_coefficients = numpy.zeros((n_qubits, n_qubits), float)
-    two_body_coefficients = numpy.zeros((n_qubits, n_qubits,
-                                         n_qubits, n_qubits), float)
-
-    # Randomly generate the one-body and two-body integrals.
-    for p in range(n_qubits):
-        for q in range(n_qubits):
-
-            # One-body terms.
-            if (p <= p) and (p % 2 == q % 2):
-                one_body_coefficients[p, q] = numpy.random.randn()
-                one_body_coefficients[q, p] = one_body_coefficients[p, q]
-
-            # Keep looping.
-            for r in range(n_qubits):
-                for s in range(n_qubits):
-
-                    # Skip zero terms.
-                    if (p == q) or (r == s):
-                        continue
-
-                    # Identify and skip one of the complex conjugates.
-                    if [p, q, r, s] != [s, r, q, p]:
-                        unique_indices = len(set([p, q, r, s]))
-
-                        # srqp srpq sprq spqr sqpr sqrp
-                        # rsqp rspq rpsq rpqs rqps rqsp.
-                        if unique_indices == 4:
-                            if min(r, s) <= min(p, q):
-                                continue
-
-                        # qqpp.
-                        elif unique_indices == 2:
-                            if q < p:
-                                continue
-
-                    # Add the two-body coefficients.
-                    two_body_coefficients[p, q, r, s] = numpy.random.randn()
-                    two_body_coefficients[s, r, q, p] = two_body_coefficients[
-                        p, q, r, s]
-
-    # Build the molecular operator and return.
-    molecular_operator = InteractionOperator(
-        constant, one_body_coefficients, two_body_coefficients)
-    return molecular_operator
 
 
 def benchmark_molecular_operator_jordan_wigner(n_qubits):
@@ -85,7 +34,7 @@ def benchmark_molecular_operator_jordan_wigner(n_qubits):
         runtime: The number of seconds required to make the conversion.
     """
     # Get an instance of InteractionOperator.
-    molecular_operator = artificial_molecular_operator(n_qubits)
+    molecular_operator = random_interaction_operator(n_qubits)
 
     # Convert to a qubit operator.
     start = time.time()
@@ -169,7 +118,7 @@ def benchmark_jordan_wigner_sparse(n_qubits):
         runtime: The time in seconds that the benchmark took.
     """
     # Initialize a random FermionOperator.
-    molecular_operator = artificial_molecular_operator(n_qubits)
+    molecular_operator = random_interaction_operator(n_qubits)
     fermion_operator = get_fermion_operator(molecular_operator)
 
     # Map to SparseOperator class.
